Add files via upload
This dataset contains around 10,000 rows of unique retailers and the amount of sales they have done for 12 brands. We have performed K-means clustering on the dataset to identify retailer segmentation. Also, after the creation of a feature called cluster class, we can presume it to be a dependent variable with 4 class labels and have saved the new dataset with dependent feature. The other dataset uploaded contains dependent feature.